Helper import_asclaras_donation - bug fix donation value - bug fix Donation first or create block
Donation model
- set value back to Float data type

+++ app/models/donation.rb
@@ -9,7 +9,7 @@ class Donation
   key :owner_id, ObjectId, :required => :true
   belongs_to :owner
 
-  key :value, BigDecimal
+  key :value, Float
   key :type, String
 
   validates_presence_of :value
